requirements.txt
numpy==1.19.3
Pillow==8.4.0
praw==7.2.0
requests==2.25.1
spacy==3.0.5
tldextract==3.1.0
tqdm==4.51.0
wikipedia==1.4.0
